Fold ToolCallingMixin into AKDAgent
- AKDAgent was the only consumer — mixin methods move onto the class directly
- _find_tool, _execute_tool, _execute_tools_parallel now live on AKDAgent
- tool_calling.py keeps ToolCall and ToolResult data models (used by adapters)
- Drop ToolCallingMixin from akd._base exports
